# fix compat mode perf_rawevents