--- a/perf/perf_rawevents.py
+++ b/perf/perf_rawevents.py
@@ -45,7 +45,10 @@ class PerfRawevents(Test):
         smm = SoftwareManager()
         detected_distro = distro.detect()
         self.distro_name = detected_distro.name
-        self.cpu_family = cpu.get_family()[5:]
+        processor = genio.read_file("/proc/cpuinfo")
+        for line in processor.splitlines():
+            if 'revision' in line:
+                self.rev = (line.split(':')[1].strip())
         if detected_distro.arch != 'ppc64le':
             self.cancel('This test is not supported on %s architecture'
                         % detected_distro.arch)
@@ -72,7 +75,11 @@ class PerfRawevents(Test):
         # Clear the dmesg to capture the delta at the end of the test.
         process.run("dmesg -C")
 
-    def run_event(self, filename, perf_flags):
+    def run_event(self, filename, eventname):
+        if eventname == 'raw':
+            perf_flags = "perf stat -e r"
+        elif eventname == 'name':
+            perf_flags = "perf stat -e "
         for line in genio.read_all_lines(filename):
             cmd = "%s%s sleep 1" % (perf_flags, line)
             output = process.run(cmd, shell=True,
@@ -86,17 +93,24 @@ class PerfRawevents(Test):
                 self.log.info("Failed command: %s", self.fail_cmd[cmd])
             self.fail("perf_raw_events: refer log file for failed events")
 
-    def test_raw_code(self):
-        file_name = 'raw_codes_p' + self.cpu_family
-        perf_flags = "perf stat -e r"
-        self.run_event(file_name, perf_flags)
-        self.error_check()
-
-    def test_name_event(self):
-        file_name = 'name_events_p' + self.cpu_family
-        perf_flags = "perf stat -e "
-        self.run_event(file_name, perf_flags)
-        self.error_check()
+    def test(self):
+        if '004b' in self.rev:
+            self.run_event('raw_codes_p8', 'raw')
+            self.error_check()
+            self.run_event('name_events_p8', 'name')
+            self.error_check()
+        if '004e' in self.rev:
+            self.run_event('raw_codes_p9', 'raw')
+            self.error_check()
+            self.run_event('name_events_p9', 'name')
+            self.error_check()
+        if '0080' in self.rev:
+            self.run_event('raw_codes_p10', 'raw')
+            self.error_check()
+            self.run_event('name_events_p10', 'name')
+            self.error_check()
+        else:
+            self.cancel('This test is not supported on %s' % cpu_family)
 
     def tearDown(self):
         # Collect the dmesg