added code to run script correctly in compat mode instead of
power model using pvr values

Signed-off-by: Disha Goel <disgoel@linux.vnet.ibm.com>

before patch:
JOB ID     : 7aa2d9c6866ab11fa4675e897e6f587b1c52d88c
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2022-06-30T11.19-7aa2d9c/job.log
 (1/2) /home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/perf_rawevents.py:PerfRawevents.test_raw_code: STARTED
 (1/2) /home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/perf_rawevents.py:PerfRawevents.test_raw_code:  PASS (1015.55 s)
 (2/2) /home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/perf_rawevents.py:PerfRawevents.test_name_event: STARTED
 (2/2) /home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/perf_rawevents.py:PerfRawevents.test_name_event:  FAIL: perf_raw_events: refer log file for failed events (133.25 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2022-06-30T11.19-7aa2d9c/results.html
JOB TIME   : 1158.93 s

after patch:
JOB ID     : d4a43c314c98450d4a916619db1488747a87d344
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2022-07-06T06.15-d4a43c3/job.log
 (1/1) perf_rawevents.py:PerfRawevents.test: STARTED
 (1/1) perf_rawevents.py:PerfRawevents.test: PASS (534.37 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2022-07-06T06.15-d4a43c3/results.html
JOB TIME   : 535.79 s

[perf_rawevents-job.log](https://github.com/avocado-framework-tests/avocado-misc-tests/files/9054753/perf_rawevents-job.log)